kursor_bot_sasha/requirements.txt
telethon==1.34.0
openai==1.3.0
python-dotenv==1.0.0
asyncio
aiohttp==3.9.1
python-telegram-bot==20.7 